+++ src/test/java/school/redrover/FolderTest.java
@@ -28,6 +28,40 @@ public class FolderTest extends BaseTest {
         getDriver().findElement(By.xpath("//a[@id = 'jenkins-home-link']")).click();
     }
 
+    private void createFolder() {
+
+        getDriver().findElement(By.linkText("Create a job")).click();
+        getDriver().findElement(By.id("name")).sendKeys(FOLDER_NAME);
+        getDriver().findElement(By.xpath("//span[@class='label' and text()='Folder']"))
+                .click();
+        getDriver().findElement(By.id("ok-button")).click();
+    }
+
+    private void creatingNewFolder (String folderName){
+        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
+        getDriver().findElement(By.className("jenkins-input")).sendKeys(folderName);
+        getDriver().findElement(By.className("com_cloudbees_hudson_plugins_folder_Folder")).click();
+        getDriver().findElement(By.xpath("//*[@id='ok-button']")).click();
+        returnToJenkinsDashboard();
+    }
+
+    private void getDashboardLink() {
+        getDriver().findElement(By.xpath("//li/a[@href='/']")).click();
+    }
+
+    private void utilsGoNameField() {
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//li[@class = 'com_cloudbees_hudson_plugins_folder_Folder']")).click();
+    }
+
+    private void create(String folderName) {
+
+        getDriver().findElement(By.linkText("New Item")).click();
+        getDriver().findElement(By.id("name")).sendKeys(folderName);
+        getDriver().findElement(By.className("com_cloudbees_hudson_plugins_folder_Folder")).click();
+        getDriver().findElement(By.id("ok-button")).click();
+        getDriver().findElement(By.name("Submit")).click();
+    }
     @Test
     public void testCreate() {
         HomePage homePage = new HomePage(getDriver())
@@ -61,7 +95,8 @@ public class FolderTest extends BaseTest {
         getDriver().findElement(By.id("ok-button")).click();
         returnToJenkinsDashboard();
 
-        getWait5().until(ExpectedConditions.elementToBeClickable(By.xpath("//td/a[@href='job/" + NESTED_FOLDER + "/']"))).click();
+        getWait5().until(ExpectedConditions.elementToBeClickable(By.xpath(
+                "//td/a[@href='job/" + NESTED_FOLDER + "/']"))).click();
         getDriver().findElement(By.xpath("//a[@href='/job/" + NESTED_FOLDER + "/move']")).click();
         getDriver().findElement(By.name("destination")).click();
         getWait5().until(ExpectedConditions.elementToBeClickable(
@@ -107,39 +142,12 @@ public class FolderTest extends BaseTest {
                 "Project " + JOB_NAME);
     }
 
-    private void createFolder() {
-
-        getDriver().findElement(By.linkText("Create a job")).click();
-        getDriver().findElement(By.id("name")).sendKeys(FOLDER_NAME);
-        getDriver().findElement(By.xpath("//span[@class='label' and text()='Folder']"))
-                .click();
-        getDriver().findElement(By.id("ok-button")).click();
-    }
-
-    private void getDashboardLink() {
-        getDriver().findElement(By.xpath("//li/a[@href='/']")).click();
-    }
-
-    private void utilsGoNameField() {
-        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
-        getDriver().findElement(By.xpath("//li[@class = 'com_cloudbees_hudson_plugins_folder_Folder']")).click();
-    }
-
-    private void create(String folderName) {
-
-        getDriver().findElement(By.linkText("New Item")).click();
-        getDriver().findElement(By.id("name")).sendKeys(folderName);
-        getDriver().findElement(By.className("com_cloudbees_hudson_plugins_folder_Folder")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-        getDriver().findElement(By.name("Submit")).click();
-    }
-
     @Test
     public void testRenameWithValidNameFromDropDownMenu() {
-
         final String folderName = "Folder1";
         final String renamedFolder = "Folder111";
         create(folderName);
+
         getDriver().findElement(By.linkText("Dashboard")).click();
 
         getDriver().findElement(By.xpath("//*[@id='job_" + folderName + "']/td[3]/a")).click();
@@ -152,121 +160,6 @@ public class FolderTest extends BaseTest {
         Assert.assertTrue(getDriver().findElement(By.xpath("//tr[@id='job_" + renamedFolder + "']")).isDisplayed());
     }
 
-    @Ignore
-    @Test
-    public void TestMoveFolder() {
-        final String firstFolderName = "Original Folder";
-        final String secondFolderName = "Inserted Folder";
-
-        create(firstFolderName);
-        getDriver().findElement(By.linkText("Dashboard")).click();
-
-        create(secondFolderName);
-        getDriver().findElement(By.linkText("Dashboard")).click();
-
-        getDriver().findElement(By.xpath("//*[@id='job_" + secondFolderName + "']/td[3]/a")).click();
-        getDriver().findElement(By.xpath("//*[@href='/job/Inserted%20Folder/move']")).click();
-        getDriver().findElement(By.xpath("//*[@id='main-panel']/form/select")).click();
-        getDriver().findElement(By.xpath("//*[@id='main-panel']/form/select/option[2]")).click();
-        getDriver().findElement(By.xpath("//*[@id='main-panel']/form/button")).click();
-
-        getDriver().findElement(By.linkText("Dashboard")).click();
-        getDriver().findElement(By.xpath("//*[@id= 'job_" + firstFolderName + "']/td[3]/a")).click();
-
-        assertEquals(getDriver().findElement(By.xpath("//*[@id='job_" + secondFolderName + "']/td[3]/a/span")).getText(), secondFolderName);
-    }
-
-    @Ignore
-    @Test
-    public void testRenameFolderUsingBreadcrumbDropdownOnFolderPage() {
-
-        final String NEW_FOLDER_NAME = "FolderNew";
-
-        createFolder();
-
-        getDriver().findElement(By.xpath("//div[@id='breadcrumbBar']//li[3]")).click();
-        getDriver().findElement(By.xpath("//a[@href='/job/" + FOLDER_NAME + "/confirm-rename']")).click();
-
-        getDriver().findElement(By.name("newName")).clear();
-        getDriver().findElement(By.name("newName")).sendKeys(NEW_FOLDER_NAME);
-        getDriver().findElement(By.name("Submit")).click();
-
-        Assert.assertEquals(getDriver().findElement(By.tagName("h1")).getText(), NEW_FOLDER_NAME,
-                FOLDER_NAME + " is not equal " + NEW_FOLDER_NAME);
-    }
-
-    @Ignore
-    @Test
-    public void testErrorMessageIsDisplayedWithoutFolderName() {
-        String expectedErrorMessage = "» This field cannot be empty, please enter a valid name";
-
-        getDriver().findElement(By.xpath("//a[@href='newJob']")).click();
-        getDriver().findElement(By.xpath("//li[@class='com_cloudbees_hudson_plugins_folder_Folder']")).click();
-        boolean errorMessageDisplayed = getDriver().findElement(By.id("itemname-required")).isDisplayed();
-        String actualErrorMessage = getDriver().findElement(By.id("itemname-required")).getText();
-
-        Assert.assertTrue(errorMessageDisplayed, "Error message for empty name is not displayed!");
-        Assert.assertEquals(actualErrorMessage, expectedErrorMessage, "The error message does not match the expected message!");
-    }
-
-    @Ignore
-    @Test
-    public void testOKbuttonIsNotClickableWithoutFolderName() {
-        getDriver().findElement(By.xpath("//a[@href='newJob']")).click();
-        getDriver().findElement(By.xpath("//li[@class='com_cloudbees_hudson_plugins_folder_Folder']")).click();
-        WebElement okButton = getDriver().findElement(By.id("ok-button"));
-        boolean okButtonDisabled = "true".equals(okButton.getAttribute("disabled"));
-
-        Assert.assertTrue(okButtonDisabled, "OK button is clickable when it shouldn't be!");
-    }
-
-    @Ignore
-    @Test
-    public void testCreatedPipelineWasBuiltSuccessfullyInCreatedFolder() {
-
-        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
-        getDriver().findElement(By.id("name")).sendKeys("Folder");
-        getDriver().findElement(By.xpath("//span[text()='Folder']")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-
-        getDriver().findElement(By.xpath("//a[@href='/job/Folder/newJob']")).click();
-        getDriver().findElement(By.id("name")).sendKeys("Pipeline");
-        getDriver().findElement(By.xpath("//span[text()='Pipeline']")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-
-        getDriver().findElement(By.xpath("//a[@href='/job/Folder/job/Pipeline/build?delay=0sec']")).click();
-
-        new Actions(getDriver())
-                .moveToElement(getDriver().findElement(By.xpath("//a[@href='/job/Folder/job/Pipeline/1/console']")))
-                .perform();
-
-
-        Assert.assertEquals(getDriver().findElement(
-                        By.xpath("//a[@href='/job/Folder/job/Pipeline/1/console']")).getAttribute("tooltip"),
-                "Success > Console Output");
-    }
-
-    @Ignore
-    @Test(dependsOnMethods = "testCreatedPipelineWasBuiltSuccessfullyInCreatedFolder")
-    public void testDeletePipelineInsideOfFolder() {
-        getDriver().findElement(By.xpath("//a[@href='job/Folder/']")).click();
-        new Actions(getDriver())
-                .moveToElement(getDriver().findElement(By.xpath("//a[@href='/job/Folder/configure']")))
-                .click()
-                .perform();
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-
-
-        getDriver().findElement(By.xpath("//a[@href='job/Pipeline/']")).click();
-        getDriver().findElement(By.xpath("//a[@data-url='/job/Folder/job/Pipeline/doDelete']")).click();
-
-        getDriver().switchTo().alert().accept();
-
-        Assert.assertEquals(getDriver().findElement(By.className("h4")).getText(), "This folder is empty");
-    }
-
     @Test
     public void testCreateNameSpecialCharacters() {
         List<String> invalidNames = Arrays.asList("#", "&", "?", "!", "@", "$", "%", "^", "*", "|", "/", "\\", "<", ">", "[", "]", ":", ";");
@@ -276,11 +169,8 @@ public class FolderTest extends BaseTest {
         WebElement inputName = getDriver().findElement(By.xpath("//input[@class = 'jenkins-input']"));
 
         for (String invalidName : invalidNames) {
-
             inputName.sendKeys(invalidName);
-
             Assert.assertEquals(getDriver().findElement(By.xpath("//div[@id = 'itemname-invalid']")).getText(), "» ‘" + invalidName + "’ is an unsafe character");
-
             inputName.clear();
         }
     }
@@ -306,9 +196,10 @@ public class FolderTest extends BaseTest {
         getDriver().findElement(By.xpath("//input[@class = 'jenkins-input']")).sendKeys(NAME_FOR_BOUNDARY_VALUES.repeat(256));
         getDriver().findElement(By.xpath("//button[@type = 'submit']")).click();
 
-        Assert.assertEquals(getDriver().findElement(By.xpath("//h2[@style = 'text-align: center']")).getText(), "A problem occurred while processing the request.");
+        Assert.assertEquals(getDriver().findElement(By.xpath(
+                "//h2[@style = 'text-align: center']")).getText(),
+                "A problem occurred while processing the request.");
     }
-
 
     @Test(dependsOnMethods = "testCreate")
     public void testAddDescriptionToFolder() {
@@ -352,6 +243,68 @@ public class FolderTest extends BaseTest {
 
         String appearanceOfAddDescriptionButton = getDriver().findElement(By.xpath("//div[@id='description']/div[2]")).getText();
         Assert.assertEquals(appearanceOfAddDescriptionButton, "Add description");
+    }
+
+    @Test
+    public void testClickPreview() {
+        createFolder();
+
+        getDriver().findElement(By.name("_.description")).sendKeys("FolderName");
+        getDriver().findElement(By.className("textarea-show-preview")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.className("textarea-preview")).getText(), "FolderName");
+    }
+
+    @Test
+    public void testNestedFolder() {
+        creatingNewFolder("FirstFolder");
+        returnToJenkinsDashboard();
+
+        creatingNewFolder("SecondFolder");
+        returnToJenkinsDashboard();
+
+        getDriver().findElement(By.xpath("//*[@id=\"job_SecondFolder\"]/td[3]/a/span")).click();
+        getDriver().findElement(By.xpath("//a[@href='/job/SecondFolder/move']")).click();
+        getDriver().findElement(By.xpath("//*[@id=\"main-panel\"]/form/select")).click();
+        getDriver().findElement(By.xpath("/html/body/div[2]/div[2]/form/select/option[2]")).click();
+        getDriver().findElement(By.xpath("//*[@id=\"main-panel\"]/form/button")).click();
+
+        returnToJenkinsDashboard();
+        Assert.assertEquals(getDriver().findElement(By.xpath("//*[@id=\"job_FirstFolder\"]/td[3]/a/span")).getText(),
+                "FirstFolder");
+
+        getDriver().findElement(By.xpath("//*[@id=\"job_FirstFolder\"]/td[3]/a/span")).click();
+        Assert.assertEquals(getDriver().findElement(By.xpath("//*[@id=\"job_SecondFolder\"]/td[3]/a/span")).getText(),
+                "SecondFolder");
+    }
+
+    @Test
+    public void testEditDescriptionFolder() {
+        creatingNewFolder("OfficeFolder");
+        getDriver().findElement(By.xpath("//span[contains(text(), 'OfficeFolder')]")).click();
+
+        getDriver().findElement(By.id("description-link")).click();
+        getDriver().findElement(By.name("description")).sendKeys("New description");
+        getDriver().findElement(By.name("Submit")).click();
+        getDriver().findElement(By.id("description-link")).click();
+        getDriver().findElement(By.name("description")).clear();
+        getDriver().findElement(By.name("description")).sendKeys("New description edited");
+        getDriver().findElement(By.name("Submit")).click();
+        Assert.assertEquals(getDriver().findElement(By.xpath(
+                "//div[@id = 'description']/div[1]")).getText(),
+                "New description edited");
+    }
+
+    @Test
+    public void testDeleteDescriptionFolder() {
+        creatingNewFolder("OfficeFolder");
+        getDriver().findElement(By.xpath("//span[contains(text(), 'OfficeFolder')]")).click();
+        getDriver().findElement(By.id("description-link")).click();
+        getDriver().findElement(By.name("description")).clear();
+        getDriver().findElement(By.name("Submit")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.cssSelector(
+                "#description > div:nth-child(1)")).getText(),"");
     }
 }
